Close xvms-943_update_logs branch
Compilation fixes...
VMS-1150: Bookmarks improvements
VMS-1150: intermediate commit
VMS-943: Add some logs to QnInstallUpdatesPeerTask

<==== Mercurial data ====>
Branch: prod_2.5

+++ mediaserver_core/src/recorder/storage_manager.cpp
@@ -1765,7 +1765,7 @@ QnStorageResourcePtr QnStorageManager::getOptimalStorageRoot(QnAbstractMediaStre
 
     struct StorageSpaceInfo {
         QnStorageResourcePtr storage;
-        qint64 usageCoeff;
+        double usageCoeff;
     };
     std::vector<StorageSpaceInfo> storagesInfo;
 
@@ -1782,7 +1782,7 @@ QnStorageResourcePtr QnStorageManager::getOptimalStorageRoot(QnAbstractMediaStre
                         .arg((*it)->getUrl())
                         .arg((*it)->calcUsageCoeff());
         if ((*it)->calcUsageCoeff() >= 0) {
-            StorageSpaceInfo tmp = {*it, static_cast<qint64>((*it)->calcUsageCoeff())};
+            StorageSpaceInfo tmp = {*it, (*it)->calcUsageCoeff()};
             storagesInfo.push_back(tmp);
         }
     }
